// frontend/apps/web/src/domains/api/generated/model/cVChunkRead.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * PostPilot
 * OpenAPI spec version: 0.1.0
 */
import type { CVChunkType } from './cVChunkType';
import type { CVChunkReadMetadata } from './cVChunkReadMetadata';
import type { CVChunkReadEmbeddingModel } from './cVChunkReadEmbeddingModel';

export interface CVChunkRead {
  id: string;
  chunk_type: CVChunkType;
  content: string;
  metadata: CVChunkReadMetadata;
  chunk_index: number;
  embedding_model: CVChunkReadEmbeddingModel;
  created_at: string;
}