// frontend/apps/web/src/domains/api/generated/model/cVChunkType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * PostPilot
 * OpenAPI spec version: 0.1.0
 */

export type CVChunkType = typeof CVChunkType[keyof typeof CVChunkType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CVChunkType = {
  summary: 'summary',
  experience: 'experience',
  skill: 'skill',
  education: 'education',
  project: 'project',
  language: 'language',
  other: 'other',
} as const;